refactor: Service가 Repository 인터페이스에 의존하도록 수정
