+++ src/main/java/roomescape/repository/JdbcReservationTimeRepository.java
@@ -1,0 +1,57 @@
+package roomescape.repository;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+import roomescape.entity.ReservationTime;
+
+import java.sql.PreparedStatement;
+import java.util.List;
+
+@Repository
+public class JdbcReservationTimeRepository implements ReservationTimeRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public JdbcReservationTimeRepository(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    public static final RowMapper<ReservationTime> reservationTimeRowMapper = (row, rowNum) ->
+            new ReservationTime(row.getLong("id"), row.getTime("start_at").toLocalTime()
+            );
+
+    @Override
+    public Long add(final ReservationTime reservationTime) {
+        String sql = "INSERT INTO reservation_time (start_at) VALUES (?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        jdbcTemplate.update(connection -> {
+            PreparedStatement ps = connection.prepareStatement(sql, new String[] {"id"});
+            ps.setString(1, String.valueOf(reservationTime.getStartAt()));
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().longValue();
+    }
+
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT * FROM reservation_time";
+        return jdbcTemplate.query(sql, reservationTimeRowMapper);
+    }
+
+    @Override
+    public void deleteById(final Long id) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+        jdbcTemplate.update(sql, id);
+    }
+
+    @Override
+    public ReservationTime findById(final Long id) {
+        String sql = "SELECT * FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, reservationTimeRowMapper, id);
+    }
+}
